requirements.txt
streamlit
azure-storage-blob
requests
pyodbc
pymssql